ditch cross-compiler and get graphics working
Turns out you don't need a cross-compiler if use just the write CFLAGS
with clang. You can get it to construct an ELF executable using the
MS_ABI, as a .so file, and construct the .efi file from the .so file
using objdump.

Graphics work now, with access to the linear framebuffer.
Now that I have the address of the framebuffer, I can call
ExitBootServices and launch the tetris application.

For now I think I'll build the tetris code directly into the EFI
application, but an interesting stretch goal would be to implement
tetris as a separate (ELF!) application and load it from disk.