main loop; broken
Drawing is too slow for us to redraw the whole image every frame, so
we only redraw when the `update` function tells us to. In tetris, many
consecutive frames can pass in which nothing changes on screen, so
this should hopefully avoid flickering.

Drawing the tetromino is broken, and just results in a black screen.